+++ !Scrabble/game/scrabble.py
@@ -1,3 +1,4 @@
 from game.tile import Tile
 from game.bag_tiles import BagTiles
 from game.player import Player
+from game.square import Square